Fix tests, vol. 1
