+++ source/core/tests/bronze/infrastructure/settings/test_submitted_transactions_stream_settings.py
@@ -8,7 +8,7 @@ def test__submitted_transactions_stream_settings__continuous_streaming_should_be
     os.environ["CONTINUOUS_STREAMING_ENABLED"] = "true"
 
     # Act
-    actual = SubmittedTransactionsStreamSettings()
+    actual = SubmittedTransactionsStreamSettings()  # type: ignore
 
     # Assert
     assert actual.continuous_streaming_enabled is True
