+++ source/core/tests/bronze/infrastructure/streams/test_kafka_stream.py
@@ -1,6 +1,6 @@
 from unittest import mock
 
-import tests.bronze.helpers.environment_variables_helpers as environment_variables_helpers
+import tests.helpers.environment_variables_helpers as environment_variables_helpers
 from core.bronze.infrastructure.streams.kafka_stream import KafkaStream
 
 
@@ -16,17 +16,3 @@ def test__submit_transactions__should_set_environment_variables(mock_get_checkpo
 
     # Act & Assert
     kafka_stream.submit_transactions(spark_mock)
-
-
-# todo: moved to silver
-@mock.patch("core.bronze.infrastructure.streams.kafka_stream.get_checkpoint_path")
-def test__write_stream__should_set_environment_variables(mock_get_checkpoint_path):
-    # Arrange
-    mock_get_checkpoint_path.return_value = "/tmp/checkpoints/start_write_stream_test"
-    environment_variables_helpers.set_kafka_authentication_settings()
-    environment_variables_helpers.set_storage_account_settings()
-    kafka_stream = KafkaStream()
-    dataframe_mock = mock.Mock()
-
-    # Act & Assert
-    kafka_stream.write_stream(dataframe_mock)
