+++ source/core/tests/helpers/environment_variables_helpers.py
