+++ source/core/tests/settings/test_storage_account_settings.py
@@ -1,6 +1,6 @@
 import os
 
-from core.bronze.infrastructure.settings import StorageAccountSettings
+from core.settings import StorageAccountSettings
 
 
 def test__submitted_transactions_stream_settings__continuous_streaming_should_be_evaluated():
@@ -9,7 +9,7 @@ def test__submitted_transactions_stream_settings__continuous_streaming_should_be
     os.environ["DATALAKE_STORAGE_ACCOUNT"] = expected_datalake_storage_account
 
     # Act
-    actual = StorageAccountSettings()
+    actual = StorageAccountSettings()  # type: ignore
 
     # Assert
     assert actual.DATALAKE_STORAGE_ACCOUNT == expected_datalake_storage_account
